billa_at: make global by adding BG, CZ and SK stores
Now returns 1818 stores instead of 1265 for just AT.

Fixes #4508
